pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "promptops-nexus2"
version = "0.1.0"
description = "Prompt versioning and system-level evals for Nexus 2.0"
requires-python = ">=3.10"
dependencies = [
  "pyyaml>=6.0",
  "requests>=2.31.0",
  "pytest>=8.0.0",
]

[tool.setuptools]
py-modules = []
